Kernel/amd64 - Clean up compile warnings

+++ Kernel/Core/arch/amd64/tss.rs
@@ -30,7 +30,7 @@ extern "C" {
 	static mut GDT: [GDTEnt; 7+MAX_CPUS*2];
 	static mut TSSes: [TSS; 4];
 	
-	static s_tid0_tls_base: usize;
+	static s_tid0_tls_base: u64;
 }
 
 fn init()
